Change batch-status to batch status cli
